test: more rust

+++ clear.ps1
@@ -1,3 +1,8 @@
-Remove-Item -Recurse -Path build
-Remove-Item -Path build.zip
-mkdir build
+if (Test-Path -Path build) 
+{
+    Remove-Item -Recurse -Path build   
+}
+if (Test-Path -Path build.zip) 
+{
+    Remove-Item -Path build.zip
+}
